codigo-fonte/backend/requirements.txt
django~=3.2.0
djangorestframework~=3.12.0